Reformat code
This attempts to make the code style consistant without making drastic
changes to the style. It was done with the astyle utility (and some by
hand) with the following options file:

lineend=linux           # (already using)
indent=tab              # Tabs for indentation, spaces for alignment
pad-oper                # 2 + 2 * 4 instead of 2+2*4
pad-header              # if () instead of if()
unpad-paren             # if (1) instead of if ( 1 )
align-reference=name    # char *foo instead of char* foo
remove-brackets         # Single-statement conditionals don't use brackets

Let me know if there is is anything that should be done differently
(such as indentation style).